requirements.txt
pydantic
pdfplumber
openai
python-docx
